src/diffusers/examples/controlnet/requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
ftfy
tensorboard
datasets
xformers==0.0.16